Move report to helper method
